Assessing-Technology-Readiness/pyproject.toml
[project]
name = "assessing-technology-readiness"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0.3",
    "questionary>=2.1.1",
    "typer>=0.20.0",
]
